moved contents of BinarySearchTree to BinaryTree
